requirements.txt
pyairtable==1.1.0
resend
assemblyai
flask
python-dotenv
openai